added tabbed activity, fixed typo
