+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,15 @@
         android:roundIcon="@drawable/time_image"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".GoalsActivity"></activity>
+        <activity android:name=".GoalsActivity" />
         <activity
             android:name=".DisplayMessageActivity"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".MainActivityOld">
 
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".MainActivityOld" />
         </activity>
         <activity android:name=".Splashscreen">
             <intent-filter>
@@ -26,14 +26,17 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity">
+        <activity
+            android:name=".MainActivity"
+            android:label="@string/title_activity_main"
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
+        <activity android:name=".MainActivityOld"></activity>
     </application>
-
 
 </manifest>
