+++ app/src/main/java/com/example/trackerczasu/DisplayMessageActivity.java
@@ -17,7 +17,7 @@ public class DisplayMessageActivity extends AppCompatActivity {
         setContentView(R.layout.activity_display_message);
         // Get the Intent that started this activity and extract the string
         Intent intent = getIntent();
-        String message = intent.getStringExtra(MainActivity.EXTRA_MESSAGE);
+        String message = intent.getStringExtra(MainActivityOld.EXTRA_MESSAGE);
 
         // Capture the layout's TextView and set the string as its text
         final TextView textView = (TextView) findViewById(R.id.textView);
