+++ app/src/main/res/drawable/makeup.png
